requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
tensorflow>=2.13.0
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=9.0.0
